refactor: extracted drop down menu to be stand alone composable
