+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/component/DropdownMenu.kt
@@ -1,0 +1,50 @@
+package com.cairosquad.evolvefit.ui.component
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
+import com.cairosquad.evolvefit.design_system.component.CheckboxItem
+import com.cairosquad.evolvefit.design_system.component.CheckboxStyle
+import com.cairosquad.evolvefit.design_system.theme.Theme
+
+@Composable
+fun DropdownMenu(
+    items: List<String>,
+    selectedItem: String?,
+    expanded: Boolean,
+    onItemClicked: (String) -> Unit,
+    onDismissRequest: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    AnimatedVisibility(
+        visible = expanded,
+        enter = fadeIn(animationSpec = tween(500)),
+        exit = fadeOut(animationSpec = tween(500))
+    ) {
+        Column(
+            modifier = modifier.fillMaxWidth().clip(RoundedCornerShape(8.dp))
+                .background(Theme.color.surfaces.surfaceContainer).padding(8.dp)
+        ) {
+            items.forEach { item ->
+                val isSelected = item == selectedItem
+
+                CheckboxItem(
+                    text = item, isChecked = isSelected, onCheckedChange = {
+                        onItemClicked(item)
+                        onDismissRequest()
+                    }, style = CheckboxStyle.Tick
+                )
+            }
+        }
+    }
+}
